+++ python/tests/integration/sdk/test_job_ops.py
@@ -1,6 +1,9 @@
 #
 # Copyright (c) 2018-2025, NVIDIA CORPORATION. All rights reserved.
 #
+
+# pylint: disable=duplicate-code
+
 from datetime import datetime, timedelta, timezone
 import unittest
 
@@ -15,7 +18,8 @@ class TestJobOps(ParallelTestBase):  # pylint: disable=unused-variable
     @pytest.mark.nonparallel("lru incompatible with some job types")
     def test_job_start_wait(self):
         job_id = self.client.job(job_kind="lru").start()
-        self.client.job(job_id=job_id).wait()
+        result = self.client.job(job_id=job_id).wait(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
         self.assertNotEqual(0, self.client.job(job_id=job_id).status().end_time)
 
     def test_job_wait_for_idle(self):
@@ -25,9 +29,10 @@ class TestJobOps(ParallelTestBase):  # pylint: disable=unused-variable
         }
         self.assertEqual(set(obj_names), existing_names)
 
-        # Start a deletion job that will reach an idle state when it finishes
+        # Start a deletion job that will reach an finished state
         job_id = self.bucket.objects(obj_names=obj_names).delete()
-        self.client.job(job_id).wait_for_idle(timeout=TEST_TIMEOUT)
+        result = self.client.job(job_id).wait(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
 
         self.assertEqual(
             0, len(self.bucket.list_objects(prefix=self.obj_prefix).entries)
@@ -37,7 +42,8 @@ class TestJobOps(ParallelTestBase):  # pylint: disable=unused-variable
         object_names = list(self._create_objects().keys())
         # Delete does not idle when finished
         job_id = self.bucket.objects(obj_names=object_names).delete()
-        self.client.job(job_id=job_id).wait(timeout=TEST_TIMEOUT)
+        result = self.client.job(job_id=job_id).wait(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
         # Check that objects do not exist
         existing_obj = {entry.name for entry in self.bucket.list_all_objects()}
         self.assertTrue(set(object_names).isdisjoint(existing_obj))
@@ -50,12 +56,14 @@ class TestJobOps(ParallelTestBase):  # pylint: disable=unused-variable
         obj, _ = self._create_object_with_content()
 
         evict_job_id = self.bucket.objects(obj_names=[obj.name]).evict()
-        self.client.job(evict_job_id).wait(timeout=TEST_TIMEOUT)
+        result = self.client.job(evict_job_id).wait(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
         self.assertFalse(obj.props.present)
 
         job_id = obj.blob_download()
         self.assertNotEqual(job_id, "")
-        self.client.job(job_id=job_id).wait_single_node(timeout=TEST_TIMEOUT)
+        result = self.client.job(job_id=job_id).wait_single_node(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
 
         self.assertTrue(obj.props.present)
 
@@ -63,7 +71,8 @@ class TestJobOps(ParallelTestBase):  # pylint: disable=unused-variable
     def test_get_within_timeframe(self):
         start_time = datetime.now(timezone.utc) - timedelta(seconds=1)
         job_id = self.client.job(job_kind="lru").start()
-        self.client.job(job_id=job_id).wait()
+        result = self.client.job(job_id=job_id).wait(timeout=TEST_TIMEOUT)
+        self.assertTrue(result.success)
         end_time = datetime.now(timezone.utc) + timedelta(seconds=1)
         self.assertNotEqual(job_id, "")
         jobs_list = self.client.job(job_id=job_id).get_within_timeframe(